MultimodalRAG/requirements.txt
pillow==10.2.0
python-dotenv==1.0.1
pathlib>=1.0.1
typing>=3.7.4
chromadb==0.4.22
tqdm==4.66.2
transformers>=4.37.0
torch>=2.0.0
mistralai==0.0.12
sentence-transformers
tiktoken
fitz
pdfplumber
frontend
pymupdf==1.23.26
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
openai>=1.0.0
easyocr>=1.7.2